Delete unused Range.h header
The functionality provided is much better off in RandomColor.h instead of a generic-named header
