+++ src/RandomColor.h
@@ -43,12 +43,35 @@
 
 #include "Color.h"
 #include "Random.h"
-#include "Range.h"
 
+#include <cstdlib>
 #include <map>
 #include <vector>
 
 namespace RandomColorGenerator {
+	// Represents a range using an upper and lower value.
+	class Range {
+	public:
+		int Lower;
+		int Upper;
+
+		Range() {}
+		Range(int lower, int upper)
+		{
+			Lower = lower;
+			Upper = upper;
+		}
+
+		// Gets the lower range for an index of 0 and the upper for an index of 1.
+		const int &operator[](const size_t index) const
+		{
+			if (index == 0)
+				return Lower;
+			else
+				return Upper;
+		}
+	};
+
 	enum ColorScheme {
 		// Select randomly from among the other color schemes.
 		SCHEME_RANDOM = 0,
